+++ gui/morpheus_model.cpp
@@ -275,6 +275,23 @@ QList<MorphModelEdit>  MorphModel::applyAutoFixes(QDomDocument document) {
 	
 	if (morpheus_file_version == morpheus_ml_version) {
 		// nothing to do ...
+		fix_version=morpheus_ml_version;
+	}
+	else if (morpheus_file_version == 3) {
+		fix_version=4;
+		MorphModel::AutoFix a;
+		a.match_path = "MorpheusModel/CellTypes/CellType/CellReporter"; a.target_path = "MorpheusModel/CellTypes/CellType/Mapper"; fixes.append(a);
+		a.match_path = "MorpheusModel/CPM/ShapeBoundary"; a.target_path = "MorpheusModel/CPM/ShapeSurface"; fixes.append(a);
+		a.match_path = "MorpheusModel/Analysis/Gnuplotter/Plot/CellLabels/@symbol-ref";a.target_path = "MorpheusModel/Analysis/Gnuplotter/Plot/CellLabels/@value"; fixes.append(a);
+		a.match_path = "MorpheusModel/Analysis/Logger/Restriction/@force-node-granularity";a.target_path = "MorpheusModel/Analysis/Logger/Input/@force-node-granularity"; fixes.append(a);
+		
+		a.operation = AutoFix::COPY; a.replace_existing = false; a.match_path  = "MorpheusModel/CellPopulations/Population/Cell/@name"; a.target_path = "MorpheusModel/CellPopulations/Population/Cell/@id"; fixes.append(a);
+		
+		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Point"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Point"; fixes.append(a);
+		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Box"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Box"; fixes.append(a);
+		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Sphere"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Sphere"; fixes.append(a);
+		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Ellipsoid"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Ellipsoid"; fixes.append(a);
+		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Cylinder"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Cylinder"; fixes.append(a);
 	}
 	else if (morpheus_file_version == 2) {
 		// return edits;
@@ -363,23 +380,10 @@ QList<MorphModelEdit>  MorphModel::applyAutoFixes(QDomDocument document) {
 	else {
 		throw  ModelException(ModelException::InvalidVersion, QString("Incompatible MorpheusML version %1").arg(morpheus_file_version) );
 	}
-	// Current language patches, will become ml_version 4
+	
+	// Current language patches
 	if (morpheus_file_version == morpheus_ml_version || fix_version == morpheus_ml_version ) {
 // 		return edits;
-		fix_version=3;
-		MorphModel::AutoFix a;
-		a.match_path = "MorpheusModel/CellTypes/CellType/CellReporter"; a.target_path = "MorpheusModel/CellTypes/CellType/Mapper"; fixes.append(a);
-		a.match_path = "MorpheusModel/CPM/ShapeBoundary"; a.target_path = "MorpheusModel/CPM/ShapeSurface"; fixes.append(a);
-		a.match_path = "MorpheusModel/Analysis/Gnuplotter/Plot/CellLabels/@symbol-ref";a.target_path = "MorpheusModel/Analysis/Gnuplotter/Plot/CellLabels/@value"; fixes.append(a);
-		a.match_path = "MorpheusModel/Analysis/Logger/Restriction/@force-node-granularity";a.target_path = "MorpheusModel/Analysis/Logger/Input/@force-node-granularity"; fixes.append(a);
-		
-		a.operation = AutoFix::COPY; a.replace_existing = false; a.match_path  = "MorpheusModel/CellPopulations/Population/Cell/@name"; a.target_path = "MorpheusModel/CellPopulations/Population/Cell/@id"; fixes.append(a);
-		
-		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Point"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Point"; fixes.append(a);
-		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Box"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Box"; fixes.append(a);
-		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Sphere"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Sphere"; fixes.append(a);
-		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Ellipsoid"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Ellipsoid"; fixes.append(a);
-		a.operation = AutoFix::MOVE; a.match_path  = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Object/Cylinder"; a.target_path = "MorpheusModel/CellPopulations/Population/InitCellObjects/Arrangement/Cylinder"; fixes.append(a);
 		// TODO Allow to specify a target path, finished by / to keep the node name!
 	}